+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning 2.0](https://semver.org/spec/v2.0.0.html).
 
 ## v2.10.1
+### Changed
+- Improved error message when logging in with a Lykos account fails due to incorrect email or password.
 ### Fixed
 - Fixed package launch not working when environment variable `SETUPTOOLS_USE_DISTUTILS` is set due to conflict with a default environment variable. User environment variables will now correctly override any default environment variables.
+- Fixed "No refresh token found" error when failing to login with Lykos account in some cases.
 
 ## v2.10.0
 ### Added